implemented follower other player mod
Hold F and Middle Click someone to start following them. It is recommended to unfollow if you want to move on your own again. Hold F and Middle Click no one to unfollow. It will also automatically unfollow if the player dies or you die. Use this feature on the side with PhantomMod to achieve more discretion