restrict possess enemy with enabled PhantomMode
Due to the possibility of accidentally teleporting enemies to the PossessionMod's game object when unpossessing enemies without ever going to phantom mode, which if possession mod is never actually run once, will be at position 0,0,0, enemies will bug out and it is undesirable for this to happen. Thus now you can no longer select an enemy to possess if no in Phantom Mode, and this problem will be fixed.